support batchnorm2d and getitem
